refactor: 5단계 - 자동차 경주(리팩토링), 코드리뷰 반영
- 패키지 구조 변경
