+++ src/main/java/racing/car/domain/Cars.java
@@ -1,4 +1,4 @@
-package carRacing.domain;
+package racing.car.domain;
 
 import java.util.ArrayList;
 import java.util.List;
