+++ src/main/java/racing/car/domain/Location.java
@@ -1,4 +1,4 @@
-package carRacing.domain;
+package racing.car.domain;
 
 import java.util.Objects;
