+++ src/main/java/racing/car/exception/CarNameException.java
@@ -1,4 +1,4 @@
-package carRacing.exception;
+package racing.car.exception;
 
 public class CarNameException extends RuntimeException {
