+++ src/main/java/racing/car/strategy/MoveStrategy.java
@@ -1,4 +1,4 @@
-package carRacing.strategy;
+package racing.car.strategy;
 
 @FunctionalInterface
 public interface MoveStrategy {
